order_service/requirements.txt
amqp==5.3.1
asgiref==3.7.2
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.7
djangorestframework==3.15.2
kombu==5.4.2
pika==1.3.2
pika-stubs==0.1.3
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
